workspacekit: Take apart rings.go
